+++ docs/conf.py
@@ -44,7 +44,7 @@ source_suffix = {
 }
 
 templates_path = ["_templates"]
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", "archive", "planning"]
 
 
 # -- Options for HTML output -------------------------------------------------